Update flavor schema tests
