+++ tests/schemas/test_flavor_schema.py
@@ -1,91 +1,22 @@
-from random import randint
-from typing import Any, Literal
-from uuid import UUID, uuid4
+from typing import Any
+from uuid import UUID
 
-import pytest
-from pytest_cases import case, parametrize, parametrize_with_cases
+from pytest_cases import parametrize_with_cases
 
 from fed_reg.flavor.models import Flavor
 from fed_reg.flavor.schemas import (
     FlavorBase,
     FlavorBasePublic,
-    FlavorCreate,
-    FlavorQuery,
     FlavorRead,
     FlavorReadPublic,
     FlavorUpdate,
 )
-from fed_reg.models import BaseNode, BaseNodeCreate, BaseNodeQuery, BaseNodeRead
 from fed_reg.provider.schemas_extended import FlavorCreateExtended
 from tests.create_dict import flavor_schema_dict
-from tests.utils import random_lower_string
 
 
-class CaseAttr:
-    @case(tags=["base_public", "base", "update"])
-    def case_none(self) -> tuple[None, None]:
-        return None, None
-
-    @case(tags=["base_public", "base"])
-    def case_desc(self) -> tuple[Literal["description"], str]:
-        return "description", random_lower_string()
-
-    @case(tags=["base"])
-    @parametrize(attr=["disk", "ram", "vcpus", "swap", "ephemeral", "gpus"])
-    def case_integer(self, attr: str) -> tuple[str, int]:
-        return attr, randint(0, 100)
-
-    @case(tags=["base"])
-    @parametrize(value=[True, False])
-    @parametrize(attr=["is_public", "infiniband"])
-    def case_boolean(self, attr: str, value: bool) -> tuple[str, bool]:
-        return attr, value
-
-    @case(tags=["base"])
-    @parametrize(attr=["gpu_model", "gpu_vendor", "local_storage"])
-    def case_string(self, attr: str) -> tuple[str, str]:
-        return attr, random_lower_string()
-
-    @case(tags=["create_extended"])
-    @parametrize(len=[0, 1, 2])
-    def case_projects(self, len: int) -> list[UUID]:
-        if len == 1:
-            return [uuid4()]
-        elif len == 2:
-            return [uuid4(), uuid4()]
-        return []
-
-
-class CaseInvalidAttr:
-    @case(tags=["base_public", "base", "update"])
-    @parametrize(attr=["name", "uuid"])
-    def case_attr(self, attr: str) -> tuple[str, None]:
-        return attr, None
-
-    @case(tags=["base"])
-    @parametrize(attr=["disk", "ram", "vcpus", "swap", "ephemeral", "gpus"])
-    def case_integer(self, attr: str) -> tuple[str, Literal[-1]]:
-        return attr, -1
-
-    @case(tags=["base"])
-    @parametrize(attr=["gpu_model", "gpu_vendor"])
-    def case_gpu_details(self, attr: str) -> tuple[str, str]:
-        return attr, random_lower_string()
-
-    @case(tags=["create_extended"])
-    @parametrize(len=[0, 1, 2])
-    def case_projects(self, len: int) -> tuple[list[UUID], str]:
-        if len == 1:
-            return [uuid4()], "Public flavors do not have linked projects"
-        elif len == 2:
-            i = uuid4()
-            return [i, i], "There are multiple identical items"
-        return [], "Projects are mandatory for private flavors"
-
-
-@parametrize_with_cases("key, value", cases=CaseAttr, has_tag=["base_public"])
+@parametrize_with_cases("key, value", has_tag="base_public")
 def test_base_public(key: str, value: str) -> None:
-    assert issubclass(FlavorBasePublic, BaseNode)
     d = flavor_schema_dict()
     if key:
         d[key] = value
@@ -95,17 +26,8 @@ def test_base_public(key: str, value: str) -> None:
     assert item.uuid == d.get("uuid").hex
 
 
-@parametrize_with_cases("key, value", cases=CaseInvalidAttr, has_tag=["base_public"])
-def test_invalid_base_public(key: str, value: None) -> None:
-    d = flavor_schema_dict()
-    d[key] = value
-    with pytest.raises(ValueError):
-        FlavorBasePublic(**d)
-
-
-@parametrize_with_cases("key, value", cases=CaseAttr, has_tag=["base"])
+@parametrize_with_cases("key, value", has_tag="base")
 def test_base(key: str, value: Any) -> None:
-    assert issubclass(FlavorBase, FlavorBasePublic)
     d = flavor_schema_dict()
     if key:
         d[key] = value
@@ -127,25 +49,8 @@ def test_base(key: str, value: Any) -> None:
     assert item.local_storage == d.get("local_storage")
 
 
-@parametrize_with_cases("key, value", cases=CaseInvalidAttr, has_tag=["base"])
-def test_invalid_base(key: str, value: Any) -> None:
-    d = flavor_schema_dict()
-    d[key] = value
-    with pytest.raises(ValueError):
-        FlavorBase(**d)
-
-
-def test_create() -> None:
-    assert issubclass(FlavorCreate, BaseNodeCreate)
-    assert issubclass(FlavorCreate, FlavorBase)
-
-
-@parametrize_with_cases(
-    "key, value", cases=[CaseInvalidAttr, CaseAttr], has_tag=["update"]
-)
+@parametrize_with_cases("key, value", has_tag="update")
 def test_update(key: str, value: Any) -> None:
-    assert issubclass(FlavorUpdate, BaseNodeCreate)
-    assert issubclass(FlavorUpdate, FlavorBase)
     d = flavor_schema_dict()
     if key:
         d[key] = value
@@ -154,13 +59,8 @@ def test_update(key: str, value: Any) -> None:
     assert item.uuid == (d.get("uuid").hex if d.get("uuid") else None)
 
 
-def test_query() -> None:
-    assert issubclass(FlavorQuery, BaseNodeQuery)
-
-
-@parametrize_with_cases("projects", cases=CaseAttr, has_tag=["create_extended"])
+@parametrize_with_cases("projects", has_tag="create_extended")
 def test_create_extended(projects: list[UUID]) -> None:
-    assert issubclass(FlavorCreateExtended, FlavorCreate)
     d = flavor_schema_dict()
     d["is_public"] = len(projects) == 0
     d["projects"] = projects
@@ -168,26 +68,8 @@ def test_create_extended(projects: list[UUID]) -> None:
     assert item.projects == [i.hex for i in projects]
 
 
-@parametrize_with_cases(
-    "projects, msg", cases=CaseInvalidAttr, has_tag=["create_extended"]
-)
-def test_invalid_create_extended(projects: list[UUID], msg: str) -> None:
-    d = flavor_schema_dict()
-    if len(projects) == 0 or len(projects) == 2:
-        d["is_public"] = False
-    elif len(projects) == 1:
-        d["is_public"] = True
-    d["projects"] = projects
-    with pytest.raises(ValueError, match=msg):
-        FlavorCreateExtended(**d)
-
-
-@parametrize_with_cases("key, value", cases=CaseAttr, has_tag=["base_public"])
+@parametrize_with_cases("key, value", has_tag="base_public")
 def test_read_public(flavor_model: Flavor, key: str, value: str) -> None:
-    assert issubclass(FlavorReadPublic, FlavorBasePublic)
-    assert issubclass(FlavorReadPublic, BaseNodeRead)
-    assert FlavorReadPublic.__config__.orm_mode
-
     if key:
         flavor_model.__setattr__(key, value)
     item = FlavorReadPublic.from_orm(flavor_model)
@@ -199,19 +81,8 @@ def test_read_public(flavor_model: Flavor, key: str, value: str) -> None:
     assert item.uuid == flavor_model.uuid
 
 
-@parametrize_with_cases("key, value", cases=CaseInvalidAttr, has_tag=["base_public"])
-def test_invalid_read_public(flavor_model: Flavor, key: str, value: str) -> None:
-    flavor_model.__setattr__(key, value)
-    with pytest.raises(ValueError):
-        FlavorReadPublic.from_orm(flavor_model)
-
-
-@parametrize_with_cases("key, value", cases=CaseAttr, has_tag=["base"])
+@parametrize_with_cases("key, value", has_tag="base")
 def test_read(flavor_model: Flavor, key: str, value: Any) -> None:
-    assert issubclass(FlavorRead, FlavorBase)
-    assert issubclass(FlavorRead, BaseNodeRead)
-    assert FlavorRead.__config__.orm_mode
-
     if key:
         flavor_model.__setattr__(key, value)
         if key.startswith("gpu_"):
@@ -234,13 +105,6 @@ def test_read(flavor_model: Flavor, key: str, value: Any) -> None:
     assert item.gpu_model == flavor_model.gpu_model
     assert item.gpu_vendor == flavor_model.gpu_vendor
     assert item.local_storage == flavor_model.local_storage
-
-
-@parametrize_with_cases("key, value", cases=CaseInvalidAttr, has_tag=["base"])
-def test_invalid_read(flavor_model: Flavor, key: str, value: str) -> None:
-    flavor_model.__setattr__(key, value)
-    with pytest.raises(ValueError):
-        FlavorRead.from_orm(flavor_model)
 
 
 # TODO Test read extended classes
